+++ tests/testthat/test-utils.R
@@ -76,3 +76,41 @@ test_that("stop_if_not_in_slide_range", {
   error_text <- tryCatch(foo(), error = paste)
   grepl("^Error:", error_text, fixed = TRUE)
 })
+
+
+test_that("update_named_defaults", {
+  opts <- options(cli.num_colors = 1) # suppress colors for error message check
+  on.exit(options(opts))
+
+  # errors
+  expect_error(update_named_defaults(x = 3, default = 1:3), "Some default vector elements have no names")
+  expect_error(update_named_defaults(x = 1:2, default = c(a = 1, b = 2, c = 3)),
+    "`x` has incorrect length (2)",
+    fixed = TRUE
+  )
+  expect_error(update_named_defaults(x = 1:2, default = c(a = 1)),
+    "Length of `x` (2) exceeds length of `default` (1)",
+    fixed = TRUE
+  )
+  expect_error(update_named_defaults(x = c(d = 3, b = 2, c = 1), default = c(aa = 1, bb = 2, cc = 3)),
+    'Found 1 unknown name in `x`: "d"',
+    fixed = TRUE
+  )
+  expect_error(update_named_defaults(x = c(b = 3, b = 2, c = 1), default = c(aa = 1, bb = 2, cc = 3)),
+    'Duplicate entries in `location`: "bb"',
+    fixed = TRUE
+  )
+  expect_error(update_named_defaults(x = c(b = 3, b = 2, c = 1), default = c(aa = 1, bb = 2, cc = 3), argname = "xxx", partial = FALSE),
+    'Found 3 unknown names in `xxx`: "b", "b", and "c"',
+    fixed = TRUE
+  )
+
+  # working
+  expect_equal(update_named_defaults(x = 2, default = c(a = 1, b = 2, c = 3)), list(a = 2, b = 2, c = 2))
+  expect_equal(update_named_defaults(x = 3:1, default = c(a = 1, b = 2, c = 3)), list(a = 3, b = 2, c = 1))
+  expect_equal(update_named_defaults(x = NULL, default = c(a = 1, b = 2, c = 3)), list(a = 1, b = 2, c = 3))
+  expect_equal(update_named_defaults(x = NULL, default = list(a = 1, b = 2, c = 3), as_list = FALSE), c(a = 1, b = 2, c = 3))
+  expect_equal(update_named_defaults(x = 3:1, default = c(a = 1, b = 2, c = 3), as_list = FALSE), c(a = 3, b = 2, c = 1))
+  expect_equal(update_named_defaults(x = c(a = 3, b = 2, c = 1), default = c(aa = 1, bb = 2, cc = 3)), list(aa = 3, bb = 2, cc = 1))
+  expect_equal(update_named_defaults(x = c(a = 99), default = c(aa = 1, bb = 2, cc = 3)), list(aa = 99, bb = 2, cc = 3))
+})